Update request.js
Wouldn't be easier on client side to have an error returned as an object rather than a mix of a text message and a json response?

I've been struggling how to handle that response, split response on "\n" and json parse the body? For some reason it doesn't work since I'm getting parsing error. Thanks.